% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_friedland_index.R
\name{calc_friedland_index}
\alias{calc_friedland_index}
\title{Calculate the Friedland index}
\usage{
calc_friedland_index(
  catch,
  chlorophyll,
  yearField = "YEAR",
  catchField = "totLand",
  chloroField = "ANNUAL_MEAN"
)
}
\arguments{
\item{catch}{Data Frame. n x anything. Data containing the total catch by each species in each year.}

\item{chlorophyll}{Data frame. n x anything.}

\item{yearField}{Character string. The name of the field in \code{catch} which contains the Yearly data.}

\item{catchField}{Character string. The name of the field in \code{catch} which contains the catch data.}

\item{chloroField}{Character string. The name of the field in \code{chlorophyll} which contains the ANNUAL_MEAN data.}
}
\value{
Data frame:

\item{YEAR}{Year of the index}
\item{INDEX}{Value of the index}
}
\description{
The Friedland index is the ratio of total catches to chlorophyll in an ecosystem.
Link & Watson (2019), "... propose a threshold of ~1 to delineate Ecosystem overfishing"
}
\details{
The units of the index: unitless
}
\section{Sources}{


J. S. Link, R. A. Watson, Global ecosystem overfishing: Clear delineation within real limits to production. Sci. Adv. 5, eaav0474 (2019). \url{DOI: 10.1126/sciadv.aav0474}
}

\seealso{
Other Link Indices: 
\code{\link{calc_fogarty_index}()},
\code{\link{calc_ryther_index}()}
}
\concept{Link Indices}
