% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean_trophic_level.R
\name{calc_mean_trophic_level}
\alias{calc_mean_trophic_level}
\title{Calculate the mean Trophic Level}
\usage{
calc_mean_trophic_level(catch, speciesTL)
}
\arguments{
\item{speciesTL}{dataframe. m x anything. Species codes and trophic level for the most abundantly caught species. One column must be named NESPP3 and another Troph}

\item{data}{dataframe. n x anything. Data containing the total catch by each species in each year. A complete grid of year by species must be passed. Columns must be named YEAR,NESPP3,catch.}
}
\value{

}
\description{
mean Trophic Level is calculated from a time series of landings: It is the sum of the species Trophic levels #' weighted by their landings
#'
}
