% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mtl_index.R
\name{calc_mtl_index}
\alias{calc_mtl_index}
\title{Calculate the mean Trophic Level}
\usage{
calc_mtl_index(
  catch,
  speciesTL,
  speciesCode = "NESPP3",
  catch_field = "totLand"
)
}
\arguments{
\item{speciesTL}{dataframe. m x anything. Species codes and trophic level for the most abundantly caught species. One column must be named NESPP3 and another Troph}

\item{speciesCode}{Character string. The name of the column that holds the speciesCodes. Default = "NESPP3".}

\item{data}{dataframe. n x anything. Data containing the total catch by each species in each year. A complete grid of year by species must be passed. Columns must be named YEAR,NESPP3,catch.}
}
\value{
Data frame:

\item{YEAR}{Year of the index}
\item{INDEX}{Value of the index}
}
\description{
mean Trophic Level is calculated from a time series of landings: It is the sum of the species Trophic levels #' weighted by their landings
#'
}
