% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PPR_scaled.R
\name{calc_ppr_scaled}
\alias{calc_ppr_scaled}
\title{Scale PPR index by total primary production}
\usage{
calc_ppr_scaled(
  PPR,
  PP,
  pprField = "INDEX",
  yearFieldPPR = "YEAR",
  ppField = "ANNUAL_MTON",
  yearFieldPP = "YEAR"
)
}
\arguments{
\item{PPR}{Data frame (n x 2). Primary production required index over time. Two columns YEAR and INDEX}

\item{PP}{List. First element labeled \code{PP} is a Data frame of size (n x 2). Columns YEAR, ANNUAL_MEAN. Second element labeled \code{EPUarea} is the area of the region}

\item{pprField}{Character string. The name of the field in \code{PPR} which contains the PPR index data.}

\item{yearFieldPPR}{Character string. The name of the field in \code{PPR} which contains the Year.}

\item{ppField}{Character string. The name of the field in \code{PP} which contains the PP data.}

\item{yearFieldPP}{Character string. The name of the field in \code{PP} which contains the Year.}
}
\value{
Data frame (n x 7)
\item{YEAR}{Year of scaled index}
\item{SCALEDINDEX}{Ratio of PPR:PP}
}
\description{
Total primary production (PP) should be in units mtC year^-1, since
PPR has units mtC year^-1.
The ratio of PPR:PP is then returned
}
\details{
Units: Percentage
}
