% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ryther_index.R
\name{calc_ryther_index}
\alias{calc_ryther_index}
\title{Calculate the Ryther index}
\usage{
calc_ryther_index(catch, area, catchField = "totLand")
}
\arguments{
\item{catch}{dataframe. n x anything. Data containg the total catch in each year. Must contain two columns labelled YEAR,catch. Can be listed by species. \code{catch} must be in metric tonnes}

\item{area}{Numeric scalar. Area of the region (km^2) in which catched come from.}

\item{catchField}{Character string. The name of the field in \code{catch} which contains the catch data.}
}
\value{
Data frame:

\item{YEAR}{Year of the index}
\item{INDEX}{Value of the index}
}
\description{
The Ryther index is composed of total catch presented on a unit area basis for an ecosystem
 using ideas found in Link & Watson (2019). See sources below.
}
\details{
Index units: metric tonnes km^-2 year^-1
}
\section{Sources}{


J. S. Link, R. A. Watson, Global ecosystem overfishing: Clear delineation within real limits to production. Sci. Adv. 5, eaav0474 (2019). \url{DOI: 10.1126/sciadv.aav0474}
}

\seealso{
Other Link Indices: 
\code{\link{calc_fogarty_index}()},
\code{\link{calc_friedland_index}()}
}
\concept{Link Indices}
