% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exlpore_species_composition.R
\name{explore_species_composition}
\alias{explore_species_composition}
\title{Explores species composition based on landings}
\usage{
explore_species_composition(channel, landings, threshold,
  filename = NULL, ...)
}
\arguments{
\item{channel}{DBI object. Connection object to database "sole"}

\item{landings}{dataframe (n x r). r >= 3. Must have 3 columns named YEAR,NESPP3,totLand}

\item{threshold}{numeric scalar. Determining the fraction of landings to be represented. Eg. threshold = 0.8
would find the fewest number of species (ordered by landings) that comprise 80\% of landings.}

\item{filename}{character string. Filename of exported figure}

\item{...}{Other arguments passed to ggplot::ggsave}
}
\value{
List:
\item{p}{ggplot object}
\item{data}{Tibble (n x 5). Columns: YEAR,NESPP3,LANDINGS,COMNAME,PresenceAbsence}
}
\description{
Species that comprise the top x% of landings for each year are extracted.
}
\section{Attention}{


For this to work, you need an oracle client installed.
Tested with Oracle instantClient_12_2 installed
Note: if you use 64 bit Rstudio then you need a 64 bit client
Note: if you use 32 bit Rstudio then you need a 32 bit client

You'll also need an account set up on the local server
}

