% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annual_PP.R
\name{get_annual_PP}
\alias{get_annual_PP}
\title{Reads primary production data}
\usage{
get_annual_PP(yrs, epu)
}
\arguments{
\item{yrs}{Numeric vector. Years to in which to extract #'@param epu character string. Name of EPU ("GB", "GOM", "MAB", "SS")}
}
\value{
list
\item{PP}{data frame. YEAR, ANNUAL_MEAN}
\item{EPUarea}{Numeric scalar. Area in Km2}
}
\description{
Kim H provided data aggregated to EPU (ststaistical areas)
}
