% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comland_data.R
\name{get_comland_data}
\alias{get_comland_data}
\title{Get landings data (COMLAND)}
\usage{
get_comland_data(channel = NULL, dataPull = F)
}
\arguments{
\item{channel}{DBI object. Connection object to database "sole"}

\item{dataPull}{Boolean. Pull from database(T) or use data from existing pull (F)}
}
\value{
data frame:

\item{YEAR}{}
\item{NESPP3}{}
\item{EPU}{}
\item{SPPLIVMT}{}
\item{US}{}
}
\description{
Pulls data from cfdbs, cleans, agregates etc. (comland script) or used a data set already pulled
(time saving step)
}
\section{First time use}{


The very first time this function in used a data pull is required to save the file to your local drive.
After which you can reference this file for future work.

The initial data pull can take 15 min
}

\section{Attention}{


#' For this to work, you need an oracle client installed.
Tested with Oracle instantClient_12_2 installed
Note: if you use 64 bit Rstudio then you need a 64 bit client
Note: if you use 32 bit Rstudio then you need a 32 bit client

You'll also need an account set up on the local server
}

