% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mtl_index.R
\name{plot_mtl_index}
\alias{plot_mtl_index}
\title{Plot the mean trophic level index}
\usage{
plot_mtl_index(MTL, epu)
}
\arguments{
\item{MTL}{data frame. (nx2). columns must be called YEAR, INDEX}

\item{epu}{character string. Name of EPU ("GB", "GOM", "MAB", "SS")}
}
\description{
Plot the mean trophic level index
}
