% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pp_index.R
\name{plot_pp_index}
\alias{plot_pp_index}
\title{Plot the Primary Production}
\usage{
plot_pp_index(scaled, epu)
}
\arguments{
\item{scaled}{data frame. (nx2). columns must be called YEAR, ANNUAL_MEAN}

\item{epu}{character string. Name of EPU ("GB", "GOM", "MAB", "SS")}
}
\description{
Plot the Primary Production
}
