% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ppr_index.R
\name{plot_ppr_index}
\alias{plot_ppr_index}
\title{Plot the Primary Production Required index (scaled)}
\usage{
plot_ppr_index(scaled, epu, withLandings = F)
}
\arguments{
\item{scaled}{data frame. (nx3). columns must be called YEAR, INDEX, landings}

\item{epu}{character string. Name of EPU ("GB", "GOM", "MAB", "SS")}

\item{withLandings}{Boolean. Create a plot with second axis, including landings ("T"). Defaul = F}
}
\description{
Plot the Primary Production Required index (scaled)
}
