% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ryther_index.R
\name{plot_ryther_index}
\alias{plot_ryther_index}
\title{Plot the Ryther index}
\usage{
plot_ryther_index(data, epu, yearField = "YEAR", indexField = "Index")
}
\arguments{
\item{data}{Data frame. columns must be called YEAR, INDEX, landings}

\item{epu}{Character string. Name of EPU ("GB", "GOM", "MAB", "SS")}

\item{yearField}{Character string. The name of the field in \code{data} which contains the Year.}

\item{indexField}{Character string. The name of the field in \code{data} which contains the value of the index}
}
\value{
ggplot Object
}
\description{
Plot the Ryther index
}
