% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_top_x_percent.R
\name{select_top_x_percent}
\alias{select_top_x_percent}
\title{Select species in top x of landings}
\usage{
select_top_x_percent(NESPP3, LANDINGS, threshold)
}
\arguments{
\item{NESPP3}{character vector. NESPP3 codes}

\item{LANDINGS}{numeric vector. landings associated with NESPP3 codes}

\item{threshold}{numeric scalar. fraction of landings used to determin the most abundant species in landings}
}
\value{
Tibble (nx2):

\item{NESPP3}{species code}
\item{LANDINGS}{landings associates with NESPP3 code}
}
\description{
A list of the most abundance species (by landings).
The number of species selected is determined by the fraction of cumulative sum
of the landings. This fraction is controlled by the value of threshold
}
