% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fogarty_index.R
\name{calc_fogarty_index}
\alias{calc_fogarty_index}
\title{Calculate the Fogarty index}
\usage{
calc_fogarty_index(
  catch,
  primaryProduction,
  yearField = "YEAR",
  catchField = "totLand",
  ppField = "ANNUAL_MEAN"
)
}
\arguments{
\item{catch}{Data frame. n x anything. Data containing the total catch by each species in each year.}

\item{primaryProduction}{Data frame. n x anything.}

\item{yearField}{Character string. The name of the field in \code{catch} which contains the Yearly data.}

\item{catchField}{Character string. The name of the field in \code{catch} which contains the catch data.}

\item{ppField}{Character string. The name of the field in \code{primaryProduction} which contains the ANNUAL_MEAN data.}
}
\value{
Data frame:

\item{YEAR}{Year of the index}
\item{INDEX}{Value of the index}
}
\description{
The Fogarty Index is the ratio of total catches to total primary productivity in an ecosystem.
Link and Watson (2019) state ".. Fogarty ration of 0.22 to  0.92 per mil, with an extreme limit of ~2.5 emerges from theoretically based limits coupled with estimates of global catches"
}
\details{
The units of the index: 0/00
}
\section{Sources}{


J. S. Link, R. A. Watson, Global ecosystem overfishing: Clear delineation within real limits to production. Sci. Adv. 5, eaav0474 (2019). \url{DOI: 10.1126/sciadv.aav0474}
}

\seealso{
Other Link Indices: 
\code{\link{calc_friedland_index}()},
\code{\link{calc_ryther_index}()}
}
\concept{Link Indices}
