% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ppr_index.R
\name{calc_ppr_index}
\alias{calc_ppr_index}
\title{Calculate the PPR index}
\usage{
calc_ppr_index(catch, speciesTL)
}
\arguments{
\item{catch}{dataframe. n x anything. Data containg the total catch by each species in each year.
Columns must be named YEAR,NESPP3,catch}

\item{speciesTL}{dataframe. m x anything. Species codes and trophic level for the most abundantly caught species.
One column must be named NESPP3 and another Troph}
}
\value{

}
\description{
Using ideas found in Pauly & Christensen's (1995) Nature publication,
Primary Production Required is calculated as:
}
