% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PPR_scaled.R
\name{calc_PPR_scaled}
\alias{calc_PPR_scaled}
\title{Scale PPR index by toal production}
\usage{
calc_PPR_scaled(PPR, PP, units = "m")
}
\arguments{
\item{PPR}{dataframe (nx2). Primary production required index over time. Two columns YEAR and INDEX}

\item{PP}{list.}
}
\value{
dataframe.
}
\description{
Scale PPR index by toal production
}
