% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ryther_index.R
\name{calc_ryther_index}
\alias{calc_ryther_index}
\title{Calculate the ryther index}
\usage{
calc_ryther_index(catch, area, catchField = "totLand")
}
\arguments{
\item{catch}{dataframe. n x anything. Data containg the total catch in each year. Must contain two columns labelled YEAR,catch. Can be listed by species. \code{catch} must be in metric tonnes}

\item{area}{Numeric scalar. Area of the region (km^2) in which catched come from.}
}
\value{
Data frame:

\item{YEAR}{Year of the index}
\item{INDEX}{Value of the index}
}
\description{
Using ideas found in Link & Watson (2019) Applied Ecology.
Index units: metric tonnes km^-2 year^-1
}
