% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ryther_index.R
\name{calc_ryther_index}
\alias{calc_ryther_index}
\title{Calculate the Ryther index}
\usage{
calc_ryther_index(catch, area, yearField = "YEAR", catchField = "totLand")
}
\arguments{
\item{catch}{Data frame. n x anything. Data containing the total catch in each year. Can be broken down by species. \code{catch} must be in metric tonnes}

\item{area}{Numeric scalar. Area of the region (km^2) in which catch comes from.}

\item{yearField}{Character string. The name of the field in \code{catch} which contains the Yearly data.}

\item{catchField}{Character string. The name of the field in \code{catch} which contains the catch data.}
}
\value{
Data frame:

\item{YEAR}{Year of the index}
\item{INDEX}{Value of the index}
}
\description{
The Ryther index is composed of total catch presented on a unit area basis for an ecosystem.
Theoretical estimates (Link & Watson, 2019), imply the index "should probably be on the order of 0.3 - 1.1 mt km^-2 y^-1 with an extreme limit of 3 mt km^-2 y^-1.
Empirically derived tipping points typically occur in fished ecosystems with total catches greater that 3 to 5 mt km^-2 y^-1 with expected system-wide MSYs on the
 order of 1 to 3 mt km^-2 y^-1"
}
\details{
The units of the index: metric tonnes km^-2 year^-1
}
\section{Sources}{


J. S. Link, R. A. Watson, Global ecosystem overfishing: Clear delineation within real limits to production. Sci. Adv. 5, eaav0474 (2019). \url{DOI: 10.1126/sciadv.aav0474}
}

\seealso{
Other Link Indices: 
\code{\link{calc_fogarty_index}()},
\code{\link{calc_friedland_index}()}
}
\concept{Link Indices}
