% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trophic_level.R
\name{get_trophic_level}
\alias{get_trophic_level}
\title{Obtain trophic level data from rfishbase}
\usage{
get_trophic_level(lookupTable)
}
\arguments{
\item{lookupTable}{dataframe. Any size. One column must be called SCIENTIFIC_NAME. Each row of the table will be cross referenced with fishbase}
}
\value{
Tibble. The same columns as \code{lookupTable} but with following additional columns.

\item{genusSpecies}{TRUE if species name is genus species, otherwise FALSE}
\item{DietTroph}{Value from DietTroph field in fishbase ecology table}
\item{FoodTroph}{Value from FoodTroph field in fishbase ecology table}
\item{vertibrate}{TRUE if species found in fishbase, FALSE if found in sealifebase, NA if not found}
}
\description{
The trophic level will be obtained for each scientific name provided.
Search will be in fishbase for vertibrates and sealifebase for invertibrates.
Two fields are accessed from the ecology table, FoodTroph and DietTroph.
See fishbase for more details (http://fishbase.us/manual/English/FishbaseThe_ECOLOGY_Table.htm)
}
